+++ src/Components/MoviesManagement/MoviesManagementContainer.js
@@ -38,35 +38,44 @@ function MoviesManagementContainer_Comp(props) {
   // check if user is authenticate (run in first render AND also when location change)
   useEffect(async () => {
 
-    // Check if User Token is valid(expired etc..)
-    let resp = await usersBL.verifyUserToken();
 
-    if (resp) {
-      if (resp.data.message == "jwt TokenExpiredError" || resp.data.message == "No token provided." || resp.data.message == "Failed to authenticate token") {
-        sessionStorage.clear();
-        showSnackbarAlret('Token is invalid or expired', 'error');
-        props.history.push('/');
+    if (CurrentUser.getPermission("View Movies")) { // Prevent appropriate permissions users to naviagte Movies page via URL
+      // Check if User Token is valid(expired etc..)
+      let resp = await usersBL.verifyUserToken();
+
+      if (resp) {
+        if (resp.data.message == "jwt TokenExpiredError" || resp.data.message == "No token provided." || resp.data.message == "Failed to authenticate token") {
+          sessionStorage.clear();
+          showSnackbarAlret('Token is invalid or expired', 'error');
+          props.history.push('/');
+        }
+        else {
+          props.history.push(url + "/allMovies");
+        }
       }
       else {
-        props.history.push(url + "/allMovies");
+        showSnackbarAlret('Token is invalid', 'error');
+        props.history.push('/');
       }
+
     }
     else {
-      showSnackbarAlret('Token is invalid', 'error');
-      props.history.push('/');
+      props.history.push('/menu');
     }
+
+
   }, []);
 
   const navigateToAllMovies = () => {
-      setAllMoviesSelected(true);
-      setAddMovieSelected(false);
-      props.history.push(url + "/allMovies");
+    setAllMoviesSelected(true);
+    setAddMovieSelected(false);
+    props.history.push(url + "/allMovies");
   }
 
   const navigateAddMovie = () => {
-      setAllMoviesSelected(false);
-      setAddMovieSelected(true);
-      props.history.push(url + "/addMovies");
+    setAllMoviesSelected(false);
+    setAddMovieSelected(true);
+    props.history.push(url + "/addMovies");
   }
 
   const showSnackbarAlret = (message, variant) => {
@@ -93,7 +102,7 @@ function MoviesManagementContainer_Comp(props) {
         <ButtonGroup disableElevation variant="contained">
           <Button color={allMoviedSelected ? "selected" : "unSelected"} value="allMovies" onClick={() => navigateToAllMovies()}>All Movies</Button>
           {/* <Button color={addMovieSelected ? "selected" : "unSelected"} value="addMovie" onClick={() => navigateAddMovie()} >Add Movie  </Button> */}
-          {CurrentUser.getPermission("Create Movies") ? <Button color={addMovieSelected ? "selected" : "unSelected"} value="addMovie" onClick={() => navigateAddMovie()} >Add Movie  </Button>  : null}
+          {CurrentUser.getPermission("Create Movies") ? <Button color={addMovieSelected ? "selected" : "unSelected"} value="addMovie" onClick={() => navigateAddMovie()} >Add Movie  </Button> : null}
         </ButtonGroup>
       </ThemeProvider>
